test: move available fields tests to compatibility test
